pyproject.toml
[project]
name = "membot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.9",
    "nltk>=3.9.1",
    "python-environ>=0.4.54",
    "sentence-transformers>=3.4.1",
    "tiktoken>=0.9.0",
]
